Split the accounting e2e lifecycle tests off the max-lines list
Break test/e2e/accounting.test.ts (1,489 lines) into a test/e2e/accounting/
folder grouped by lifecycle stage, and drop the file's entry from the
noExcessiveLinesPerFile ignore list.

The single "e2e: accounting lifecycle" describeWithEnv and its ~30 shared
helpers move to two support modules: ledger-helpers.ts (the ledger-truth
queries and rendered-figure assertions) and drivers.ts (the Stripe-success,
refund, merge and balance drivers, plus a describeAccounting() wrapper each
shard reuses). The 21 tests split across income-and-corrections,
refunds-and-conservation, free-and-idempotency, and bulk-merges-and-pwyw.

Every original test carries over unchanged; the folder's 25 nodes pass, and
the linter and duplication checker are clean.
