--- test/e2e/accounting.test.ts
@@ -1,1489 +1,0 @@
-/**
- * End-to-end money/accounting lifecycle, exercised through the REAL routes:
- * a public Stripe payment completion, the admin pages that render the resulting
- * figures, the manual money-correction routes (decision 14), and the admin
- * refund route. Every money step asserts BOTH the ledger truth (via the
- * `#shared/accounting` query functions) AND the rendered admin value (GET the
- * admin route, assert the `formatCurrency` string appears) — the two must agree.
- *
- * These are full-lifecycle flows. The narrower unit-ish coverage of the same
- * surfaces lives in test/lib/server-balance.test.ts, server-admin-balance.test.ts,
- * server-refunds.test.ts, and refund-ledger.test.ts; this file reuses their
- * helpers rather than duplicating them.
- *
- * Sign reminder (mapBooking's contract): `balanceOf(attendee)` is the NEGATIVE of
- * what the attendee owes, so OWED = −accountBalance(attendeeAccount(id)). A
- * fully-paid booking nets the attendee to 0; a listing's revenue account is the
- * destination of its `sale` legs, so its balance is the recognised income; WORLD
- * is the source of every `payment`, so taking £50 leaves WORLD at −5000. A manual
- * correction posts an `adjustment` leg against WRITEOFF and must never move WORLD.
- */
-
-import { expect } from "@std/expect";
-import { afterEach, it as test } from "@std/testing/bdd";
-import { type Stub, stub } from "@std/testing/mock";
-import { handleRequest } from "#routes";
-import {
-  attendeeAccount,
-  BOOKING_FEE_INCOME,
-  modifierAccount,
-  revenueAccount,
-  WORLD,
-  WRITEOFF,
-} from "#shared/accounting/accounts.ts";
-import {
-  MANUAL_ATTENDEE_CHARGE,
-  MANUAL_ATTENDEE_WRITEOFF,
-} from "#shared/accounting/manual-entries.ts";
-import {
-  accountBalance,
-  allTransfers,
-  transfersByAccount,
-  transfersByEventGroup,
-} from "#shared/accounting/queries.ts";
-import { formatCurrency } from "#shared/currency.ts";
-import { settleAttendeeBalance } from "#shared/db/attendees/balance.ts";
-import { getAttendeesRaw } from "#shared/db/attendees.ts";
-import { settings } from "#shared/db/settings.ts";
-import { allBalances } from "#shared/ledger/project.ts";
-import type { Transfer } from "#shared/ledger/types.ts";
-import { paymentsApi } from "#shared/payments.ts";
-import { resetStripeClient, stripeApi } from "#shared/stripe.ts";
-import { stripePaymentProvider } from "#shared/stripe-provider.ts";
-import {
-  adminFormPost,
-  adminGet,
-  createPaidTestAttendee,
-  createTestAttendee,
-  createTestAttendeeDirect,
-  createTestListing,
-  describeWithEnv,
-  expectFlash,
-  expectFlashRedirect,
-  expectHtmlResponse,
-  expectRedirect,
-  followRedirect,
-  mockProviderType,
-  mockRequest,
-  setupStripe,
-  signMeta,
-  singleItem,
-  testCookie,
-  testCsrfToken,
-  withMocks,
-} from "#test-utils";
-import { extractInputValue } from "#test-utils/csrf.ts";
-import { postListingSale, postModifierLeg } from "#test-utils/ledger.ts";
-import { mockFormRequest } from "#test-utils/mocks.ts";
-import { insertModifier } from "#test-utils/modifiers.ts";
-
-// -- Ledger-truth helpers ------------------------------------------------- //
-
-/** Normalise a signed-zero to a plain zero so `toBe(0)` (strict, `-0 !== 0`)
- *  never trips on the `-0` that negating an empty balance can produce. */
-const norm = (value: number): number => value + 0;
-
-/** A listing's recognised income (the raw `revenue` account balance — gross sale
- *  credits, less write-off AND refund debits). */
-const incomeOf = async (listingId: number): Promise<number> =>
-  norm(await accountBalance(revenueAccount(listingId)));
-
-/** What an attendee still owes — the NEGATIVE of their account balance. */
-const owedBy = async (attendeeId: number): Promise<number> =>
-  norm(-(await accountBalance(attendeeAccount(attendeeId))));
-
-/** A modifier's net revenue (positive surcharge collected, negative discount). */
-const modifierRevenueOf = async (modifierId: number): Promise<number> =>
-  norm(await accountBalance(modifierAccount(modifierId)));
-
-/** The world/cash account balance — every honest cash report reads `world→*`,
- *  so a correction must leave this untouched. */
-const worldBalance = async (): Promise<number> =>
-  norm(await accountBalance(WORLD));
-
-/** The signed sum of balances across EVERY account that has any transfer.
- *  Double-entry conservation: this is exactly 0 after any sequence of legs. */
-const sumOfAllBalances = async (): Promise<number> => {
-  const balances = allBalances(await allTransfers());
-  let total = 0;
-  for (const value of balances.values()) total += value;
-  return norm(total);
-};
-
-// -- Admin page assertions ------------------------------------------------ //
-
-/** GET an owner page and return its HTML, asserting a 200. */
-const adminPageHtml = async (path: string): Promise<string> => {
-  const response = await adminGet(path);
-  expect(response.status).toBe(200);
-  return response.text();
-};
-
-/**
- * Assert a `revenue` account's RUNNING BALANCE on the per-account ledger
- * statement page renders the given minor-unit figure (`Balance: £X`). This is
- * the raw signed balance, so a refund's `revenue→attendee` debit DOES reduce it
- * (and it can go negative once a write-off and a refund both apply).
- */
-const assertStatementBalance = async (
-  listingId: number,
-  minor: number,
-): Promise<void> => {
-  const statement = await adminPageHtml(`/admin/ledger/revenue/${listingId}`);
-  expect(statement).toContain(`Balance: ${formatCurrency(minor)}`);
-};
-
-/**
- * Assert the listing EDIT page's "Current income" input renders the given
- * minor-unit figure (`value="£X"`). The edit page shows GROSS credits minus only
- * manual write-offs (`creditsLessWriteoffDebits`), so — unlike the statement —
- * an ordinary refund does NOT reduce it (matching the legacy `SUM(price_paid)`),
- * while a manual write-off does. The two surfaces therefore agree only when no
- * refund has touched the account.
- */
-const assertEditPageIncome = async (
-  listingId: number,
-  minor: number,
-): Promise<void> => {
-  const edit = await adminPageHtml(`/admin/listing/${listingId}/edit`);
-  expect(edit).toContain(`value="${formatCurrency(minor)}"`);
-};
-
-/** With no refund applied, both income surfaces agree on the same figure. */
-const assertRenderedIncome = async (
-  listingId: number,
-  minor: number,
-): Promise<void> => {
-  await assertStatementBalance(listingId, minor);
-  await assertEditPageIncome(listingId, minor);
-};
-
-/**
- * Assert an attendee's outstanding balance, on BOTH the per-account ledger
- * statement (`Balance: £X`, where owed = −running) and the admin ledger page
- * (`Balance outstanding:` label followed by the formatted figure).
- */
-const assertRenderedOwed = async (
-  attendeeId: number,
-  minor: number,
-): Promise<void> => {
-  const formatted = formatCurrency(minor);
-  const balancePage = await adminPageHtml(
-    `/admin/attendees/${attendeeId}/ledger`,
-  );
-  expect(balancePage).toContain("Balance outstanding:");
-  expect(balancePage).toContain(formatted);
-};
-
-/**
- * Assert a modifier's revenue, on BOTH the modifier edit page (a disabled
- * `value="£X"` input) and the modifier list (a Revenue cell of the same figure).
- */
-const assertRenderedModifierRevenue = async (
-  modifierId: number,
-  minor: number,
-): Promise<void> => {
-  const formatted = formatCurrency(minor);
-  const edit = await adminPageHtml(`/admin/modifiers/${modifierId}/edit`);
-  expect(edit).toContain(`value="${formatted}"`);
-  const list = await adminPageHtml("/admin/modifiers");
-  expect(list).toContain(formatted);
-};
-
-/** The breakdown template's signed-magnitude format (a leading +/− with a U+2212
- *  minus, a plain unsigned figure for zero), replicated so a test can assert the
- *  exact reconciliation figures the page renders. */
-const signedCurrency = (value: number): string =>
-  value === 0
-    ? formatCurrency(0)
-    : `${value < 0 ? "−" : "+"}${formatCurrency(Math.abs(value))}`;
-
-/** Slice the `#income-ledger` reconciliation article out of a listing detail
- *  page, so a figure is asserted WITHIN the breakdown and can't accidentally
- *  match an unrelated figure elsewhere on the page. */
-const incomeLedgerArticle = (html: string): string => {
-  const start = html.indexOf('id="income-ledger"');
-  expect(start).toBeGreaterThan(-1);
-  return html.slice(start, html.indexOf("</article>", start));
-};
-
-/** GET the merge preview for `targetId` loaded with `sourceToken`, returning the
- *  `merge_version` the apply POST must echo back AND the name of the conflicting
- *  booking's decision field (`booking_<listingId>:<startAt>`) scraped from the
- *  rendered form, so the test answers the exact conflict the operator is shown
- *  rather than guessing the key. Uses the stable owner cookie — the preview
- *  decrypts the source's PII, needing the session's private key. */
-const mergePreview = async (
-  targetId: number,
-  sourceToken: string,
-): Promise<{ version: string; bookingField: string }> => {
-  const page = await adminGet(
-    `/admin/attendees/${targetId}/actions?token=${encodeURIComponent(
-      sourceToken,
-    )}`,
-  );
-  expect(page.status).toBe(200);
-  const html = await page.text();
-  const version = extractInputValue(html, "merge_version");
-  expect(version).not.toBeNull();
-  const bookingField = html.match(/name="(booking_[^"]+)"/)?.[1];
-  expect(bookingField).toBeDefined();
-  return { bookingField: bookingField!, version: version! };
-};
-
-/** POST the merge apply form on the SAME stable owner cookie as
- *  {@link mergePreview}, so the apply decrypts the source under the same session
- *  that built the diff (the merge needs the owner's private key). */
-const mergePost = async (
-  targetId: number,
-  fields: Record<string, string>,
-): Promise<Response> => {
-  const csrf = await testCsrfToken();
-  const cookie = await testCookie();
-  return handleRequest(
-    mockFormRequest(
-      `/admin/attendees/${targetId}/merge`,
-      { csrf_token: csrf, ...fields },
-      cookie,
-    ),
-  );
-};
-
-/** Build a listing with two fully-PAID duplicate bookings (a target and a
- *  token-bearing source) on it — the same-listing conflict decision 17 must
- *  resolve. Income counts BOTH £50 tickets (£100) until the merge un-bills the
- *  discarded one. Returns the ids plus the source's merge token. */
-const twoPaidDuplicates = async (
-  name: string,
-): Promise<{
-  listingId: number;
-  targetId: number;
-  sourceId: number;
-  sourceToken: string;
-}> => {
-  const listing = await createTestListing({
-    maxAttendees: 10,
-    name,
-    unitPrice: 5000,
-  });
-  const { attendee: target } = await createTestAttendeeDirect(
-    listing.id,
-    `${name} Target`,
-    `target-${name}@example.com`,
-  );
-  const { attendee: source, token: sourceToken } =
-    await createTestAttendeeDirect(
-      listing.id,
-      `${name} Source`,
-      `source-${name}@example.com`,
-    );
-  await postListingSale({
-    attendeeId: target.id,
-    gross: 5000,
-    listingId: listing.id,
-  });
-  await postListingSale({
-    attendeeId: source.id,
-    gross: 5000,
-    listingId: listing.id,
-  });
-  return {
-    listingId: listing.id,
-    sourceId: source.id,
-    sourceToken,
-    targetId: target.id,
-  };
-};
-
-/** The money-decision form field paired with a scraped `booking_<key>` field. */
-const moneyFieldFor = (bookingField: string): string =>
-  bookingField.replace("booking_", "money_");
-
-// -- Public-payment driver (mirrors server-payments-success.test.ts) ------ //
-
-/** A compact modifier ref as it rides the signed metadata (`{ i: id, q: qty }`). */
-type ModRef = { i: number; q: number };
-
-type StripeOrder = {
-  items: string;
-  total: number;
-  modifiers?: ModRef[];
-  name?: string;
-  email?: string;
-  sessionId?: string;
-  paymentIntent?: string;
-};
-
-/**
- * Stub the Stripe session retrieval for `order` (metadata signed exactly as
- * production — the order may span several listing lines and carry applied
- * `modifiers`) and run `body` with the REAL `/payment/success` response, then
- * restore the stub. The signed `total` MUST equal the order the handler
- * re-derives (gross + fee + modifiers) — the price oracle refunds a mismatch.
- */
-const withStripeSuccess = async (
-  order: StripeOrder,
-  body: (response: Response) => Promise<void>,
-): Promise<void> => {
-  const sessionId = order.sessionId ?? "cs_e2e";
-  const metadata = signMeta(
-    {
-      email: order.email ?? "customer@example.com",
-      items: order.items,
-      name: order.name ?? "Customer",
-      ...(order.modifiers
-        ? { modifiers: JSON.stringify(order.modifiers) }
-        : {}),
-    },
-    order.total,
-  );
-  const mockRetrieve = stub(stripeApi, "retrieveCheckoutSession", () =>
-    Promise.resolve({
-      amount_total: order.total,
-      id: sessionId,
-      metadata,
-      payment_intent: order.paymentIntent ?? "pi_e2e",
-      payment_status: "paid",
-    } as unknown as Awaited<
-      ReturnType<typeof stripeApi.retrieveCheckoutSession>
-    >),
-  );
-  try {
-    await body(
-      await handleRequest(
-        mockRequest(`/payment/success?session_id=${sessionId}`),
-      ),
-    );
-  } finally {
-    mockRetrieve.restore();
-  }
-};
-
-/** Drive a first-time Stripe success and assert the production thank-you
- *  redirect (the common case over {@link withStripeSuccess}). */
-const runStripeSuccess = (order: StripeOrder): Promise<void> =>
-  withStripeSuccess(order, async (redirect) => {
-    expectRedirect(redirect, /^\/payment\/success\?tokens=.+$/);
-    await expectHtmlResponse(
-      await followRedirect(redirect, handleRequest),
-      200,
-      "Thank you for your order",
-    );
-  });
-
-/**
- * Drive a genuine single-listing Stripe success for `gross` minor units and
- * return the attendee the booking created — the common case over
- * {@link runStripeSuccess}.
- */
-const completePaidOrder = async (
-  listingId: number,
-  name: string,
-  email: string,
-  gross: number,
-  sessionId = "cs_e2e",
-  paymentIntent = "pi_e2e",
-): Promise<number> => {
-  await runStripeSuccess({
-    email,
-    items: singleItem(listingId, 1, gross),
-    name,
-    paymentIntent,
-    sessionId,
-    total: gross,
-  });
-  const attendees = await getAttendeesRaw(listingId);
-  expect(attendees.length).toBe(1);
-  return attendees[0]!.id;
-};
-
-// -- Refund driver (mirrors server-refunds.test.ts withRefundMock) -------- //
-
-/** Run `body` with the payment provider resolved to a stripe provider whose
- *  `refundPayment` is stubbed, so the admin refund routes reach the ledger
- *  reversal without a real network call. `refund` is either a fixed outcome or a
- *  per-`paymentId` function — the latter lets a bulk refund decline one specific
- *  payment while the rest succeed. */
-const withRefundMock = (
-  refund: boolean | ((paymentId: string) => Promise<boolean>),
-  body: (mockRefund: Stub) => Promise<void>,
-): Promise<void> =>
-  withMocks(
-    () =>
-      stub(paymentsApi, "getConfiguredProvider", () =>
-        mockProviderType("stripe"),
-      ),
-    async () => {
-      const behave =
-        typeof refund === "function" ? refund : () => Promise.resolve(refund);
-      const mockRefund = stub(stripePaymentProvider, "refundPayment", behave);
-      try {
-        await body(mockRefund);
-      } finally {
-        mockRefund.restore();
-      }
-    },
-  );
-
-/** POST the real single-attendee admin refund form as the owner. */
-const submitRefund = async (
-  attendeeId: number,
-  confirmName: string,
-): Promise<Response> => {
-  const { response } = await adminFormPost(
-    `/admin/attendees/${attendeeId}/refund`,
-    { confirm_identifier: confirmName },
-  );
-  return response;
-};
-
-// -- Attendee-edit driver (scrapes the real edit form) -------------------- //
-
-/** Extract the hidden/select fields the attendee edit form round-trips
- *  (`line_listing_*`, `qty_*`, `line_key_*`, `line_package_*`, `status_id`)
- *  from the rendered edit page, so a balance correction re-submits the EXACT
- *  booking and changes only the owed figure — exactly what a browser would
- *  post back. */
-/** Move an attendee's owed balance through the ledger — the proper path now the
- *  attendee form no longer edits balances. `MANUAL_ATTENDEE_WRITEOFF` lowers what
- *  they owe (a goodwill write-off), `MANUAL_ATTENDEE_CHARGE` raises it, each by
- *  `amountMajor` pounds. Returns the route's response. */
-const postAttendeeBalanceEntry = async (
-  attendeeId: number,
-  entryType: string,
-  amountMajor: string,
-): Promise<Response> =>
-  (
-    await adminFormPost(`/admin/ledger/attendee/${attendeeId}/add`, {
-      amount: amountMajor,
-      entry_type: entryType,
-      occurred_at: "2026-06-22T12:00",
-      return_url: `/admin/attendees/${attendeeId}/edit`,
-    })
-  ).response;
-
-// -- Leg helpers ---------------------------------------------------------- //
-
-const kindsOf = (legs: Transfer[]): string[] =>
-  legs.map((leg) => leg.kind ?? "").sort();
-
-const legsOfKind = (legs: Transfer[], kind: string): Transfer[] =>
-  legs.filter((leg) => leg.kind === kind);
-
-describeWithEnv("e2e: accounting lifecycle", { db: true }, () => {
-  afterEach(() => resetStripeClient());
-
-  // 1. A genuine public paid order recognises income, leaves the buyer owing
-  //    nothing, and both admin surfaces render the income the ledger holds.
-  test("a real public paid order recognises income shown on the admin pages", async () => {
-    await setupStripe();
-    const listing = await createTestListing({
-      maxAttendees: 50,
-      name: "Workshop",
-      unitPrice: 5000,
-    });
-
-    const attendeeId = await completePaidOrder(
-      listing.id,
-      "Customer",
-      "customer@example.com",
-      5000,
-    );
-
-    // Ledger truth: gross income recognised, buyer paid in full (owes nothing).
-    expect(await incomeOf(listing.id)).toBe(5000);
-    expect(await owedBy(attendeeId)).toBe(0);
-
-    // The booking's legs are a sale + a payment under ONE event group (the
-    // booking fee defaults to 0 in a fresh setup, so there is no fee leg).
-    const legs = await transfersByAccount(attendeeAccount(attendeeId));
-    expect(kindsOf(legs)).toEqual(["payment", "sale"]);
-    expect(new Set(legs.map((leg) => leg.eventGroup)).size).toBe(1);
-    expect(legsOfKind(legs, "sale")[0]!.amount).toBe(5000);
-    expect(legsOfKind(legs, "payment")[0]!.amount).toBe(5000);
-    expect(await accountBalance(BOOKING_FEE_INCOME)).toBe(0);
-
-    // Rendered admin value agrees on both the ledger statement and the edit page.
-    await assertRenderedIncome(listing.id, 5000);
-  });
-
-  // 2. A deposit leaves the remainder owed (ledger + balance page), and settling
-  //    through the production settle clears it to zero on both.
-  test("a deposit owes the remainder until settled, on the ledger and balance page", async () => {
-    const listing = await createTestListing({
-      maxAttendees: 50,
-      name: "Retreat",
-      unitPrice: 8000,
-    });
-    // Owe the full £80 with nothing paid, then post a £30 deposit payment, so
-    // £50 remains. createTestAttendee on a provider-less priced listing already
-    // posts the gross owed sale, so the deposit is the only extra leg needed.
-    const attendee = await createTestAttendee(
-      listing.id,
-      listing.slug,
-      "Deposit Payer",
-      "deposit@example.com",
-    );
-    expect(await owedBy(attendee.id)).toBe(8000);
-    await postListingSale({
-      amountPaid: 3000,
-      attendeeId: attendee.id,
-      // gross 0 adds no new sale; only the £30 deposit payment leg is posted.
-      gross: 0,
-      listingId: listing.id,
-    });
-
-    // Ledger + admin ledger page agree: £50 outstanding.
-    expect(await owedBy(attendee.id)).toBe(5000);
-    await assertRenderedOwed(attendee.id, 5000);
-
-    // Settle the remaining balance the production way and re-check both.
-    const result = await settleAttendeeBalance(attendee.id, 5000, {
-      id: "settle-e2e",
-      occurredAt: "2026-06-22T00:00:00.000Z",
-    });
-    expect(result.settled).toBe(true);
-    expect(await owedBy(attendee.id)).toBe(0);
-    const settledPage = await adminPageHtml(
-      `/admin/attendees/${attendee.id}/ledger`,
-    );
-    expect(settledPage).toContain("This booking is fully paid");
-  });
-
-  // 3. A manual income write-off (decision 14) lowers recognised income by
-  //    exactly the delta, posts a single writeoff↔revenue adjustment, and leaves
-  //    WORLD (the cash report) untouched. A later refund still behaves sanely.
-  test("a manual income write-off lowers income without touching cash", async () => {
-    await setupStripe();
-    const listing = await createTestListing({
-      maxAttendees: 50,
-      name: "Gala",
-      unitPrice: 5000,
-    });
-    const attendeeId = await completePaidOrder(
-      listing.id,
-      "Gala Guest",
-      "gala@example.com",
-      5000,
-      "cs_gala",
-      "pi_gala",
-    );
-    expect(await incomeOf(listing.id)).toBe(5000);
-    const worldBefore = await worldBalance();
-    const writeoffBefore = await accountBalance(WRITEOFF);
-
-    // Write the recognised income down from £50 to £20 (income field is MAJOR).
-    const response = (
-      await adminFormPost(`/admin/listing/${listing.id}/income`, {
-        income: "20.00",
-      })
-    ).response;
-    await expectFlashRedirect(
-      `/admin/listing/${listing.id}/edit`,
-      "Listing income adjusted",
-    )(response);
-
-    // Income dropped by exactly £30; the correction is one writeoff↔revenue leg.
-    expect(await incomeOf(listing.id)).toBe(2000);
-    const revenueLegs = await transfersByAccount(revenueAccount(listing.id));
-    const adjustments = legsOfKind(revenueLegs, "adjustment");
-    expect(adjustments.length).toBe(1);
-    expect(adjustments[0]!.amount).toBe(3000);
-    // The write-down debits revenue toward the writeoff contra account.
-    expect(adjustments[0]!.source).toEqual(revenueAccount(listing.id));
-    expect(adjustments[0]!.destination).toEqual(WRITEOFF);
-    // WORLD is untouched (cash report stays honest); writeoff absorbed the £30.
-    expect(await worldBalance()).toBe(worldBefore);
-    expect(await accountBalance(WRITEOFF)).toBe(writeoffBefore + 3000);
-
-    // Rendered admin value reflects the corrected income on both surfaces.
-    await assertRenderedIncome(listing.id, 2000);
-
-    // A refund after a write-down still behaves sanely. The refund reverses only
-    // the booking's OWN legs (the £50 sale + £50 payment), not the manual
-    // write-off adjustment — so the buyer ends owing nothing and their cash
-    // returns to the world, but the £30 write-off remains a standing debit on the
-    // revenue account. The raw revenue balance is therefore £50 − £30 − £50 = −£30
-    // (the ledger statement shows this), while the edit page (gross-minus-write-
-    // offs, refund-agnostic) still shows £50 − £30 = £20. Conservation still holds.
-    await withRefundMock(true, async (mockRefund) => {
-      const refund = await submitRefund(attendeeId, "Gala Guest");
-      expectRedirect(refund, /\/admin\/attendees\/\d+\/actions/);
-      expect(mockRefund.calls.length).toBe(1);
-    });
-    expect(await owedBy(attendeeId)).toBe(0);
-    expect(await incomeOf(listing.id)).toBe(-3000);
-    expect(await sumOfAllBalances()).toBe(0);
-    const refundCash = legsOfKind(
-      await transfersByAccount(attendeeAccount(attendeeId)),
-      "refund_cash",
-    );
-    expect(refundCash.length).toBe(1);
-    expect(refundCash[0]!.amount).toBe(5000);
-    // Both rendered income surfaces, each against its own (divergent) contract.
-    await assertStatementBalance(listing.id, -3000);
-    await assertEditPageIncome(listing.id, 2000);
-  });
-
-  // 4. A manual attendee-balance correction moves what's owed up and then down
-  //    by exactly the delta on both the ledger and the balance page, and never
-  //    moves WORLD.
-  test("a manual attendee-balance correction moves owed up and down, cash untouched", async () => {
-    const listing = await createTestListing({
-      maxAttendees: 50,
-      name: "Series Pass",
-      unitPrice: 6000,
-    });
-    const attendee = await createTestAttendee(
-      listing.id,
-      listing.slug,
-      "Balance Edith",
-      "edith@example.com",
-    );
-    // Provider-less priced booking starts owing the full £60.
-    expect(await owedBy(attendee.id)).toBe(6000);
-    const worldStart = await worldBalance();
-
-    // Correct owed DOWN by £35 through the ledger (goodwill write-off): £60 → £25.
-    const down = await postAttendeeBalanceEntry(
-      attendee.id,
-      MANUAL_ATTENDEE_WRITEOFF,
-      "35.00",
-    );
-    expect(down.status).toBe(302);
-    expect(await owedBy(attendee.id)).toBe(2500);
-    await assertRenderedOwed(attendee.id, 2500);
-    expect(await worldBalance()).toBe(worldStart);
-
-    // Correct owed UP by £15 (a manual charge): £25 → £40.
-    const up = await postAttendeeBalanceEntry(
-      attendee.id,
-      MANUAL_ATTENDEE_CHARGE,
-      "15.00",
-    );
-    expect(up.status).toBe(302);
-    expect(await owedBy(attendee.id)).toBe(4000);
-    await assertRenderedOwed(attendee.id, 4000);
-    // Cash never moved through either correction — only the writeoff contra did.
-    expect(await worldBalance()).toBe(worldStart);
-    // The recognised sale income is unchanged; corrections touch only the
-    // attendee's clearing account against writeoff.
-    expect(await incomeOf(listing.id)).toBe(6000);
-  });
-
-  // 5. A manual modifier-revenue correction moves a seeded modifier's revenue to
-  //    the target on the ledger and on the modifier edit/list pages.
-  test("a manual modifier-revenue correction moves revenue to the target", async () => {
-    const modifier = await insertModifier({
-      calcValue: 700,
-      name: "VIP Surcharge",
-    });
-    // Seed a real surcharge leg: +£7 collected.
-    await postModifierLeg({ delta: 700, modifierId: modifier.id });
-    expect(await modifierRevenueOf(modifier.id)).toBe(700);
-    const worldBefore = await worldBalance();
-
-    // Correct the net revenue to £12 (total_revenue field is MAJOR units).
-    const response = (
-      await adminFormPost(`/admin/modifiers/${modifier.id}/revenue`, {
-        total_revenue: "12.00",
-      })
-    ).response;
-    await expectFlashRedirect(
-      `/admin/modifiers/${modifier.id}/edit`,
-      "Modifier revenue adjusted",
-    )(response);
-
-    // Ledger moved to exactly the target via a single writeoff↔modifier leg.
-    expect(await modifierRevenueOf(modifier.id)).toBe(1200);
-    const adjustments = legsOfKind(
-      await transfersByAccount(modifierAccount(modifier.id)),
-      "adjustment",
-    );
-    expect(adjustments.length).toBe(1);
-    expect(adjustments[0]!.amount).toBe(500);
-    // Raising revenue credits the modifier from writeoff.
-    expect(adjustments[0]!.source).toEqual(WRITEOFF);
-    expect(adjustments[0]!.destination).toEqual(modifierAccount(modifier.id));
-    expect(await worldBalance()).toBe(worldBefore);
-
-    // Rendered admin value agrees on the edit page and the list.
-    await assertRenderedModifierRevenue(modifier.id, 1200);
-  });
-
-  // 6. Refunding a real paid order reverses revenue→0 and owed→0, posts a full
-  //    refund_cash leg, conservation holds, and the admin listing page renders
-  //    the refunded attendee's state.
-  test("refunding a paid order returns revenue and owed to zero with conservation", async () => {
-    await setupStripe();
-    const listing = await createTestListing({
-      maxAttendees: 50,
-      name: "Concert",
-      unitPrice: 4500,
-    });
-    const attendeeId = await completePaidOrder(
-      listing.id,
-      "Refundee",
-      "refundee@example.com",
-      4500,
-      "cs_concert",
-      "pi_concert",
-    );
-    expect(await incomeOf(listing.id)).toBe(4500);
-
-    await withRefundMock(true, async (mockRefund) => {
-      const response = await submitRefund(attendeeId, "Refundee");
-      await expectFlashRedirect(
-        `/admin/attendees/${attendeeId}/actions`,
-        "Refund issued",
-      )(response);
-      expect(mockRefund.calls.length).toBe(1);
-    });
-
-    // Ledger truth: revenue and owed both back to zero.
-    expect(await incomeOf(listing.id)).toBe(0);
-    expect(await owedBy(attendeeId)).toBe(0);
-
-    // A single full refund_cash leg of the whole payment, returned to the world.
-    const refundCash = legsOfKind(
-      await transfersByAccount(attendeeAccount(attendeeId)),
-      "refund_cash",
-    );
-    expect(refundCash.length).toBe(1);
-    expect(refundCash[0]!.amount).toBe(4500);
-    expect(refundCash[0]!.destination).toEqual(WORLD);
-
-    // Conservation across every touched account.
-    expect(await sumOfAllBalances()).toBe(0);
-
-    // The two income surfaces legitimately DIVERGE after a refund: the ledger
-    // statement nets the refund (`Balance: £0`), while the edit page reports
-    // gross-minus-write-offs and so still shows the £45 sale (an ordinary refund
-    // doesn't reduce recognised income — only a manual write-off does). Assert
-    // each surface against its own contract rather than forcing them to agree.
-    await assertStatementBalance(listing.id, 0);
-    await assertEditPageIncome(listing.id, 4500);
-    // The admin attendee balance page shows the booking fully settled.
-    const balancePage = await adminPageHtml(
-      `/admin/attendees/${attendeeId}/ledger`,
-    );
-    expect(balancePage).toContain("This booking is fully paid");
-  });
-
-  // 7. Conservation sweep over a MIXED sequence: a paid order, a manual income
-  //    write-off, and a refund. The signed sum of balances across every touched
-  //    account must be exactly 0.
-  test("conservation holds across a mixed order + correction + refund sequence", async () => {
-    await setupStripe();
-    const listing = await createTestListing({
-      maxAttendees: 50,
-      name: "Festival",
-      unitPrice: 7000,
-    });
-
-    // Order #1: a paid order that we will later write down and refund.
-    const refundedId = await completePaidOrder(
-      listing.id,
-      "Mixed One",
-      "mixed1@example.com",
-      7000,
-      "cs_mix1",
-      "pi_mix1",
-    );
-    // Order #2 stays on the books, so the sweep spans accounts left non-zero.
-    await createPaidTestAttendee(
-      listing.id,
-      "Mixed Two",
-      "mixed2@example.com",
-      "pi_mix2",
-      7000,
-    );
-
-    // A manual income write-down (£70 → £40 across the two £70 sales = £140
-    // recognised; drop to £100).
-    expect(await incomeOf(listing.id)).toBe(14000);
-    const writeDown = (
-      await adminFormPost(`/admin/listing/${listing.id}/income`, {
-        income: "100.00",
-      })
-    ).response;
-    await expectFlashRedirect(
-      `/admin/listing/${listing.id}/edit`,
-      "Listing income adjusted",
-    )(writeDown);
-    expect(await incomeOf(listing.id)).toBe(10000);
-
-    // Refund order #1.
-    await withRefundMock(true, async (mockRefund) => {
-      const response = await submitRefund(refundedId, "Mixed One");
-      expectRedirect(response, /\/admin\/attendees\/\d+\/actions/);
-      expect(mockRefund.calls.length).toBe(1);
-    });
-
-    // Conservation must hold after the whole mixed sequence.
-    expect(await sumOfAllBalances()).toBe(0);
-
-    // And the surviving figures are individually coherent: order #2 still owes
-    // nothing, the refunded buyer owes nothing.
-    expect(await owedBy(refundedId)).toBe(0);
-    // The raw revenue balance (what the ledger statement shows) nets everything:
-    // £140 gross − £40 write-down − £70 refunded sale = £30.
-    expect(await incomeOf(listing.id)).toBe(3000);
-    await assertStatementBalance(listing.id, 3000);
-    // The edit page reports gross-minus-write-offs and ignores the refund, so it
-    // still shows £140 − £40 = £100 — the documented divergence after a refund.
-    await assertEditPageIncome(listing.id, 10000);
-
-    // The refund leg group is distinct from the booking group it reverses.
-    const refundCash = legsOfKind(
-      await transfersByAccount(attendeeAccount(refundedId)),
-      "refund_cash",
-    );
-    expect(refundCash.length).toBe(1);
-    const bookingGroups = new Set(
-      (await transfersByEventGroup(refundCash[0]!.eventGroup)).map(
-        (leg) => leg.eventGroup,
-      ),
-    );
-    expect(bookingGroups.size).toBe(1);
-  });
-
-  // 8. A configured booking fee posts a separate `fee` leg to fee-income (not the
-  //    listing's revenue), leaves the buyer owing nothing, and a refund reverses
-  //    that fee alongside the sale and the cash.
-  test("a configured booking fee posts fee-income and is reversed on refund", async () => {
-    await setupStripe();
-    await settings.update.bookingFee("10"); // 10% booking fee.
-    const listing = await createTestListing({
-      maxAttendees: 50,
-      name: "Fee Day",
-      unitPrice: 5000,
-    });
-    // £50 ticket + 10% booking fee = £55 charged.
-    await runStripeSuccess({
-      email: "fee@example.com",
-      items: singleItem(listing.id, 1, 5000),
-      name: "Fee Payer",
-      paymentIntent: "pi_fee",
-      sessionId: "cs_fee",
-      total: 5500,
-    });
-    const attendeeId = (await getAttendeesRaw(listing.id))[0]!.id;
-
-    // Recognised income is the gross ticket sale; the fee is its own income line.
-    expect(await incomeOf(listing.id)).toBe(5000);
-    expect(await owedBy(attendeeId)).toBe(0);
-    expect(await accountBalance(BOOKING_FEE_INCOME)).toBe(500);
-    expect(await worldBalance()).toBe(-5500);
-
-    const legs = await transfersByAccount(attendeeAccount(attendeeId));
-    expect(kindsOf(legs)).toEqual(["fee", "payment", "sale"]);
-    const fee = legsOfKind(legs, "fee")[0]!;
-    expect(fee.amount).toBe(500);
-    expect(fee.destination).toEqual(BOOKING_FEE_INCOME);
-
-    // Refunding reverses sale + fee + payment; fee income returns to zero.
-    await withRefundMock(true, async (mockRefund) => {
-      const refund = await submitRefund(attendeeId, "Fee Payer");
-      expectRedirect(refund, /\/admin\/attendees\/\d+\/actions/);
-      expect(mockRefund.calls.length).toBe(1);
-    });
-    expect(await incomeOf(listing.id)).toBe(0);
-    expect(await accountBalance(BOOKING_FEE_INCOME)).toBe(0);
-    expect(await owedBy(attendeeId)).toBe(0);
-    expect(await worldBalance()).toBe(0);
-    expect(await sumOfAllBalances()).toBe(0);
-    const refundFee = legsOfKind(
-      await transfersByAccount(attendeeAccount(attendeeId)),
-      "refund_fee",
-    );
-    expect(refundFee.length).toBe(1);
-    expect(refundFee[0]!.amount).toBe(500);
-  });
-
-  // 9. A surcharge modifier applied during a REAL paid checkout posts a `modifier`
-  //    leg whose balance is the modifier's revenue (rendered on the admin pages),
-  //    and a refund reverses it with a `refund_modifier` leg.
-  test("a surcharge modifier in a real checkout earns revenue and reverses on refund", async () => {
-    await setupStripe();
-    const listing = await createTestListing({
-      maxAttendees: 50,
-      name: "Talk",
-      unitPrice: 5000,
-    });
-    const modifier = await insertModifier({
-      calcKind: "percent",
-      calcValue: 10,
-      name: "Service charge",
-    });
-    // £50 ticket + 10% service charge = £55 charged.
-    await runStripeSuccess({
-      email: "svc@example.com",
-      items: singleItem(listing.id, 1, 5000),
-      modifiers: [{ i: modifier.id, q: 1 }],
-      name: "Svc Buyer",
-      paymentIntent: "pi_svc",
-      sessionId: "cs_svc",
-      total: 5500,
-    });
-    const attendeeId = (await getAttendeesRaw(listing.id))[0]!.id;
-
-    expect(await incomeOf(listing.id)).toBe(5000);
-    expect(await modifierRevenueOf(modifier.id)).toBe(500);
-    expect(await owedBy(attendeeId)).toBe(0);
-
-    const legs = await transfersByAccount(attendeeAccount(attendeeId));
-    expect(kindsOf(legs)).toEqual(["modifier", "payment", "sale"]);
-    const mod = legsOfKind(legs, "modifier")[0]!;
-    expect(mod.amount).toBe(500);
-    expect(mod.destination).toEqual(modifierAccount(modifier.id));
-
-    // The earned revenue renders on the modifier edit page and the list.
-    await assertRenderedModifierRevenue(modifier.id, 500);
-
-    // Refund reverses the modifier leg too, returning its revenue to zero.
-    await withRefundMock(true, async (mockRefund) => {
-      const refund = await submitRefund(attendeeId, "Svc Buyer");
-      expectRedirect(refund, /\/admin\/attendees\/\d+\/actions/);
-      expect(mockRefund.calls.length).toBe(1);
-    });
-    expect(await modifierRevenueOf(modifier.id)).toBe(0);
-    expect(await sumOfAllBalances()).toBe(0);
-    const refundMod = legsOfKind(
-      await transfersByAccount(attendeeAccount(attendeeId)),
-      "refund_modifier",
-    );
-    expect(refundMod.length).toBe(1);
-    expect(refundMod[0]!.amount).toBe(500);
-  });
-
-  // 10. One payment spanning two listings splits the recognised income across each
-  //     listing's own revenue account under a single event group, and the buyer
-  //     (one attendee) owes nothing.
-  test("a multi-line order splits income across each listing's revenue account", async () => {
-    await setupStripe();
-    const first = await createTestListing({
-      maxAttendees: 50,
-      name: "Part One",
-      unitPrice: 3000,
-    });
-    const second = await createTestListing({
-      maxAttendees: 50,
-      name: "Part Two",
-      unitPrice: 2000,
-    });
-    // One £50 payment: £30 to the first listing, £20 to the second.
-    await runStripeSuccess({
-      email: "both@example.com",
-      items: JSON.stringify([
-        { e: first.id, p: 3000, q: 1 },
-        { e: second.id, p: 2000, q: 1 },
-      ]),
-      name: "Both Buyer",
-      paymentIntent: "pi_multi",
-      sessionId: "cs_multi",
-      total: 5000,
-    });
-
-    // Each listing's revenue account holds its own line; the buyer owes nothing.
-    expect(await incomeOf(first.id)).toBe(3000);
-    expect(await incomeOf(second.id)).toBe(2000);
-    const attendeeId = (await getAttendeesRaw(first.id))[0]!.id;
-    expect(await owedBy(attendeeId)).toBe(0);
-    expect(await worldBalance()).toBe(-5000);
-    expect(await sumOfAllBalances()).toBe(0);
-
-    // Both sale legs and the single payment share ONE booking event group.
-    const legs = await transfersByAccount(attendeeAccount(attendeeId));
-    expect(legsOfKind(legs, "sale").length).toBe(2);
-    expect(new Set(legs.map((leg) => leg.eventGroup)).size).toBe(1);
-
-    // Each listing's income renders on its own edit page.
-    await assertEditPageIncome(first.id, 3000);
-    await assertEditPageIncome(second.id, 2000);
-  });
-
-  // 11. A genuinely free (£0) booking through the public ticket form records no
-  //     money at all — no sale, no payment, and (even with a booking fee
-  //     configured) no phantom fee income or cash. The attendee owes nothing.
-  test("a free booking records no money even with a booking fee configured", async () => {
-    await settings.update.bookingFee("10");
-    const listing = await createTestListing({
-      maxAttendees: 50,
-      name: "Free Meetup",
-      unitPrice: 0,
-    });
-    const attendee = await createTestAttendee(
-      listing.id,
-      listing.slug,
-      "Free Guest",
-      "free@example.com",
-    );
-
-    // Nothing owed, and the booking posted no ledger legs whatsoever.
-    expect(await owedBy(attendee.id)).toBe(0);
-    expect(
-      (await transfersByAccount(attendeeAccount(attendee.id))).length,
-    ).toBe(0);
-    expect(await incomeOf(listing.id)).toBe(0);
-    // No phantom booking-fee income, and no phantom cash moved through the world.
-    expect(await accountBalance(BOOKING_FEE_INCOME)).toBe(0);
-    expect(await worldBalance()).toBe(0);
-    expect(await sumOfAllBalances()).toBe(0);
-  });
-
-  // 12. When the payment provider DECLINES the refund, nothing is reversed: the
-  //     income and owed figures are unchanged and no refund legs are posted.
-  test("a failed provider refund reverses nothing in the ledger", async () => {
-    await setupStripe();
-    const listing = await createTestListing({
-      maxAttendees: 50,
-      name: "Show",
-      unitPrice: 4500,
-    });
-    const attendeeId = await completePaidOrder(
-      listing.id,
-      "No Refund",
-      "norefund@example.com",
-      4500,
-      "cs_fail",
-      "pi_fail",
-    );
-    expect(await incomeOf(listing.id)).toBe(4500);
-
-    await withRefundMock(false, async (mockRefund) => {
-      const response = await submitRefund(attendeeId, "No Refund");
-      // The route surfaces the failure as a flash error and does NOT issue it.
-      await expectFlashRedirect(
-        `/admin/attendees/${attendeeId}/refund`,
-        expect.stringContaining("Refund failed"),
-        false,
-      )(response);
-      expect(mockRefund.calls.length).toBe(1);
-    });
-
-    // The ledger is exactly as it was: full income, nothing owed back, no refund.
-    expect(await incomeOf(listing.id)).toBe(4500);
-    expect(await owedBy(attendeeId)).toBe(0);
-    const legs = await transfersByAccount(attendeeAccount(attendeeId));
-    expect(legsOfKind(legs, "refund_cash").length).toBe(0);
-    expect(kindsOf(legs)).toEqual(["payment", "sale"]);
-    expect(await sumOfAllBalances()).toBe(0);
-  });
-
-  // 13. Corrections and webhook deliveries are idempotent. Replaying the identical
-  //     Stripe success creates no second booking or duplicate legs, and
-  //     re-submitting the same income target posts no second adjustment.
-  test("re-submitting a correction and replaying a success are both no-ops", async () => {
-    await setupStripe();
-    const listing = await createTestListing({
-      maxAttendees: 50,
-      name: "Repeat",
-      unitPrice: 6000,
-    });
-    const attendeeId = await completePaidOrder(
-      listing.id,
-      "Repeat Buyer",
-      "repeat@example.com",
-      6000,
-      "cs_repeat",
-      "pi_repeat",
-    );
-
-    // Replaying the identical success must not create a second attendee or
-    // duplicate the booking's legs — an already-processed session is a no-op,
-    // so the route just re-renders (a 200) rather than redirecting afresh.
-    await withStripeSuccess(
-      {
-        email: "repeat@example.com",
-        items: singleItem(listing.id, 1, 6000),
-        name: "Repeat Buyer",
-        paymentIntent: "pi_repeat",
-        sessionId: "cs_repeat",
-        total: 6000,
-      },
-      async (replay) => {
-        await replay.body?.cancel();
-      },
-    );
-    expect((await getAttendeesRaw(listing.id)).length).toBe(1);
-    expect(
-      kindsOf(await transfersByAccount(attendeeAccount(attendeeId))),
-    ).toEqual(["payment", "sale"]);
-
-    // Adjust income to £40, then submit the SAME target again — the second submit
-    // computes a zero delta and posts no second adjustment.
-    const adjustIncome = async (): Promise<Response> =>
-      (
-        await adminFormPost(`/admin/listing/${listing.id}/income`, {
-          income: "40.00",
-        })
-      ).response;
-    const expectAdjusted = expectFlashRedirect(
-      `/admin/listing/${listing.id}/edit`,
-      "Listing income adjusted",
-    );
-    await expectAdjusted(await adjustIncome());
-    await expectAdjusted(await adjustIncome());
-
-    expect(await incomeOf(listing.id)).toBe(4000);
-    const adjustments = legsOfKind(
-      await transfersByAccount(revenueAccount(listing.id)),
-      "adjustment",
-    );
-    expect(adjustments.length).toBe(1);
-  });
-
-  // 14. The listing detail page renders the "Income & ledger" reconciliation that
-  //     EXPLAINS the two income figures: gross sales (+) and manual adjustments
-  //     (±) make up recognised income, and refunds (−) take it down to the net
-  //     ledger balance — so the two never silently disagree (transparency).
-  test("the listing page renders an income/ledger breakdown reconciling the figures", async () => {
-    await setupStripe();
-    const listing = await createTestListing({
-      maxAttendees: 50,
-      name: "Reconciled",
-      unitPrice: 5000,
-    });
-    const attendeeId = await completePaidOrder(
-      listing.id,
-      "Recon Buyer",
-      "recon@example.com",
-      5000,
-      "cs_recon",
-      "pi_recon",
-    );
-    // Write recognised income down £50 → £40 (a manual adjustment), then refund
-    // the booking (which nets the ledger balance but not recognised income) so
-    // all five reconciliation rows carry distinct, non-zero figures.
-    await expectFlashRedirect(
-      `/admin/listing/${listing.id}/edit`,
-      "Listing income adjusted",
-    )(
-      (
-        await adminFormPost(`/admin/listing/${listing.id}/income`, {
-          income: "40.00",
-        })
-      ).response,
-    );
-    await withRefundMock(true, async (mockRefund) => {
-      const refund = await submitRefund(attendeeId, "Recon Buyer");
-      expectRedirect(refund, /\/admin\/attendees\/\d+\/actions/);
-      expect(mockRefund.calls.length).toBe(1);
-    });
-
-    const article = incomeLedgerArticle(
-      await adminPageHtml(`/admin/listing/${listing.id}`),
-    );
-    // Every reconciliation row is labelled and signed, so the difference between
-    // the two income figures is self-evident on the page.
-    expect(article).toContain("Income &amp; ledger");
-    expect(article).toContain("Gross ticket sales");
-    expect(article).toContain(signedCurrency(5000)); // +£50 gross sales
-    expect(article).toContain("Manual adjustments");
-    expect(article).toContain(signedCurrency(-1000)); // −£10 write-down
-    expect(article).toContain("Recognised income");
-    expect(article).toContain(formatCurrency(4000)); // £40 recognised
-    expect(article).toContain("Refunds");
-    expect(article).toContain(signedCurrency(-5000)); // −£50 refunded
-    expect(article).toContain("Net balance in ledger");
-    expect(article).toContain(formatCurrency(-1000)); // −£10 net ledger balance
-    // And it links through to the listing-scoped ledger view.
-    expect(article).toContain(`/admin/ledger?listing=${listing.id}`);
-  });
-
-  // 15. A bulk refund is resilient: when the provider declines ONE payment, the
-  //     others are still refunded and recorded, the declined one is left intact,
-  //     and conservation holds across the partial batch.
-  test("a bulk refund continues past a failure and reverses only the successes", async () => {
-    await setupStripe();
-    const listing = await createTestListing({
-      maxAttendees: 50,
-      name: "Tour",
-      unitPrice: 5000,
-    });
-    const one = await createPaidTestAttendee(
-      listing.id,
-      "One",
-      "one@example.com",
-      "pi_b1",
-      5000,
-    );
-    const two = await createPaidTestAttendee(
-      listing.id,
-      "Two",
-      "two@example.com",
-      "pi_b2",
-      5000,
-    );
-    const three = await createPaidTestAttendee(
-      listing.id,
-      "Three",
-      "three@example.com",
-      "pi_b3",
-      5000,
-    );
-    expect(await incomeOf(listing.id)).toBe(15000);
-
-    // The provider declines the middle payment only.
-    await withRefundMock(
-      (paymentId: string) => Promise.resolve(paymentId !== "pi_b2"),
-      async (mockRefund) => {
-        const { response } = await adminFormPost(
-          `/admin/listing/${listing.id}/refund-all`,
-          { confirm_identifier: listing.name },
-        );
-        await expectFlashRedirect(
-          `/admin/listing/${listing.id}/refund-all`,
-          expect.stringContaining("1 failed"),
-          false,
-        )(response);
-        expect(mockRefund.calls.length).toBe(3); // all three attempted
-      },
-    );
-
-    // The two successes were reversed; the declined one keeps its sale and cash.
-    const refundCashCount = async (id: number): Promise<number> =>
-      legsOfKind(await transfersByAccount(attendeeAccount(id)), "refund_cash")
-        .length;
-    expect(await refundCashCount(one.id)).toBe(1);
-    expect(await refundCashCount(three.id)).toBe(1);
-    expect(await refundCashCount(two.id)).toBe(0);
-    // Income reflects exactly the one surviving sale; conservation still holds.
-    expect(await incomeOf(listing.id)).toBe(5000);
-    expect(await sumOfAllBalances()).toBe(0);
-  });
-
-  // 16. A refund is written to the attendee's own activity log (the audit trail
-  //     that makes a ledger change visible — transparency), and a SECOND refund is
-  //     rejected before the provider is even called, posting no duplicate leg.
-  test("a refund is logged on the attendee and a second refund is a safe no-op", async () => {
-    await setupStripe();
-    const listing = await createTestListing({
-      maxAttendees: 50,
-      name: "Logged",
-      unitPrice: 4500,
-    });
-    const attendeeId = await completePaidOrder(
-      listing.id,
-      "Logged Guest",
-      "logged@example.com",
-      4500,
-      "cs_log",
-      "pi_log",
-    );
-
-    await withRefundMock(true, async (mockRefund) => {
-      const refund = await submitRefund(attendeeId, "Logged Guest");
-      await expectFlashRedirect(
-        `/admin/attendees/${attendeeId}/actions`,
-        "Refund issued",
-      )(refund);
-      expect(mockRefund.calls.length).toBe(1);
-    });
-    expect(await owedBy(attendeeId)).toBe(0);
-    const refundCash = legsOfKind(
-      await transfersByAccount(attendeeAccount(attendeeId)),
-      "refund_cash",
-    );
-    expect(refundCash.length).toBe(1);
-
-    // Transparency: the money event shows in the attendee's activity log.
-    const activityPage = await adminPageHtml(
-      `/admin/attendees/${attendeeId}/activity`,
-    );
-    expect(activityPage).toContain("Refund issued for attendee 'Logged Guest'");
-
-    // Idempotency: a second refund is refused (already refunded) without calling
-    // the provider, and no second reversal is posted.
-    await withRefundMock(true, async (mockRefund) => {
-      const again = await submitRefund(attendeeId, "Logged Guest");
-      await expectFlashRedirect(
-        `/admin/attendees/${attendeeId}/refund`,
-        expect.stringContaining("already been refunded"),
-        false,
-      )(again);
-      expect(mockRefund.calls.length).toBe(0);
-    });
-    expect(
-      legsOfKind(
-        await transfersByAccount(attendeeAccount(attendeeId)),
-        "refund_cash",
-      ).length,
-    ).toBe(1);
-    expect(await sumOfAllBalances()).toBe(0);
-  });
-
-  // 17. Merging two PAID duplicate bookings, CREDIT choice (decision 17 — an
-  //     explicit operator choice, never a silent default): the survivor keeps ONE
-  //     ticket, the discarded duplicate's recognised sale is un-billed (so the
-  //     listing's income counts the kept ticket once, not twice), and the £50 the
-  //     duplicate paid is handed back as the survivor's CREDIT. The source account
-  //     empties, conservation holds, and the credit is shown on the balance page.
-  test("a merge crediting a paid duplicate shows the survivor's credit", async () => {
-    const { listingId, targetId, sourceId, sourceToken } =
-      await twoPaidDuplicates("Reunion");
-    // Both tickets recognised: income counts £100 before the merge resolves it.
-    expect(await incomeOf(listingId)).toBe(10000);
-    expect(await owedBy(targetId)).toBe(0);
-    expect(await owedBy(sourceId)).toBe(0);
-
-    // Keep the target ticket; CREDIT the discarded source payment (both decisions
-    // are scraped from the rendered form, then applied).
-    const { version, bookingField } = await mergePreview(targetId, sourceToken);
-    const merged = await mergePost(targetId, {
-      [bookingField]: "keep_target",
-      [moneyFieldFor(bookingField)]: "credit",
-      merge_version: version,
-      source_token: sourceToken,
-    });
-    expect(merged.status).toBe(302);
-
-    // Income now counts the ONE kept ticket; the survivor holds the over-paid £50
-    // as a credit (negative owed); the source account is emptied; conservation holds.
-    expect(await incomeOf(listingId)).toBe(5000);
-    expect(await owedBy(targetId)).toBe(-5000);
-    expect(norm(await accountBalance(attendeeAccount(sourceId)))).toBe(0);
-    expect(await sumOfAllBalances()).toBe(0);
-
-    // Transparency: the survivor's ledger page renders the £50 credit figure.
-    const balancePage = await adminPageHtml(
-      `/admin/attendees/${targetId}/ledger`,
-    );
-    expect(balancePage).toContain(formatCurrency(-5000));
-  });
-
-  // 18. The same paid duplicate, WRITE-OFF choice: income still counts the one kept
-  //     ticket, but the over-paid £50 is parked in the `writeoff` contra account
-  //     rather than credited — the survivor owes nothing and cash reports stay
-  //     honest. Conservation still holds.
-  test("a merge writing off a paid duplicate parks the cash in writeoff", async () => {
-    const { listingId, targetId, sourceId, sourceToken } =
-      await twoPaidDuplicates("Gala");
-    const writeoffBefore = norm(await accountBalance(WRITEOFF));
-
-    const { version, bookingField } = await mergePreview(targetId, sourceToken);
-    const merged = await mergePost(targetId, {
-      [bookingField]: "keep_target",
-      [moneyFieldFor(bookingField)]: "writeoff",
-      merge_version: version,
-      source_token: sourceToken,
-    });
-    expect(merged.status).toBe(302);
-
-    // One ticket's income survives; the survivor owes nothing; the un-billed £50
-    // lands in writeoff (not returned); the source empties; conservation holds.
-    expect(await incomeOf(listingId)).toBe(5000);
-    expect(await owedBy(targetId)).toBe(0);
-    expect(norm(await accountBalance(WRITEOFF))).toBe(writeoffBefore + 5000);
-    expect(norm(await accountBalance(attendeeAccount(sourceId)))).toBe(0);
-    expect(await sumOfAllBalances()).toBe(0);
-  });
-
-  // 19. The decision-17 GUARD: a paid duplicate may NOT be discarded without an
-  //     explicit money choice. Omitting it REFUSES the merge (the form re-renders
-  //     with the error instead of redirecting) and mutates NOTHING — income still
-  //     counts both tickets and the source still exists — so no silent double-count
-  //     or stranded money can slip through.
-  test("a merge refuses to discard a paid booking with no money decision", async () => {
-    const { listingId, sourceToken, targetId } =
-      await twoPaidDuplicates("Summit");
-
-    const { version, bookingField } = await mergePreview(targetId, sourceToken);
-    // Submit the booking choice but OMIT the money decision.
-    const refused = await mergePost(targetId, {
-      [bookingField]: "keep_target",
-      merge_version: version,
-      source_token: sourceToken,
-    });
-    // Bounced back to the Actions tab's merge panel with the validation
-    // error flashed; the apply never ran.
-    expect(refused.status).toBe(302);
-    expectFlash(refused, expect.stringContaining("money decision"), false);
-
-    // Nothing changed: both tickets' income still counts and both attendees survive.
-    expect(await incomeOf(listingId)).toBe(10000);
-    expect((await getAttendeesRaw(listingId)).length).toBe(2);
-    expect(await sumOfAllBalances()).toBe(0);
-  });
-
-  // 20. A FREE duplicate (no money on either side) needs NO money decision: the
-  //     preview surfaces the conflicting booking row but hides the credit/write-off
-  //     UI, and the merge applies straight through posting no reversal legs — the
-  //     "nothing at stake" path decision 17 must keep one-click.
-  test("a free duplicate merges with no money decision and no reversal legs", async () => {
-    const listing = await createTestListing({
-      maxAttendees: 10,
-      name: "Freebie",
-      unitPrice: 0,
-    });
-    const { attendee: target } = await createTestAttendeeDirect(
-      listing.id,
-      "Free A",
-      "free-a@example.com",
-    );
-    const { token: sourceToken } = await createTestAttendeeDirect(
-      listing.id,
-      "Free B",
-      "free-b@example.com",
-    );
-    // Nothing paid on either side — £0 at stake.
-    expect(await incomeOf(listing.id)).toBe(0);
-
-    // The preview shows the conflict row but NOT the money-decision UI.
-    const preview = await adminGet(
-      `/admin/attendees/${target.id}/actions?token=${encodeURIComponent(
-        sourceToken,
-      )}`,
-    );
-    const previewHtml = await preview.text();
-    expect(previewHtml).toContain('name="booking_');
-    expect(previewHtml).not.toContain("merge-money-decision");
-    expect(previewHtml).not.toContain("Discarded payment");
-
-    const version = extractInputValue(previewHtml, "merge_version")!;
-    const bookingField = previewHtml.match(/name="(booking_[^"]+)"/)![1]!;
-    const merged = await mergePost(target.id, {
-      [bookingField]: "keep_target",
-      merge_version: version,
-      source_token: sourceToken,
-    });
-    // No money decision required, so the merge redirects straight through.
-    expect(merged.status).toBe(302);
-
-    // The merge went through with no money moved and the source folded in.
-    expect(await incomeOf(listing.id)).toBe(0);
-    expect((await getAttendeesRaw(listing.id)).length).toBe(1);
-    expect(await sumOfAllBalances()).toBe(0);
-  });
-
-  // 21. A pay-what-you-want (can_pay_more) order recognises the FULL chosen price
-  //     as income — the figure shown is exactly what the buyer paid, not the base
-  //     price — and they owe nothing.
-  test("a pay-what-you-want order recognises the chosen price as income", async () => {
-    await setupStripe();
-    const listing = await createTestListing({
-      canPayMore: true,
-      maxAttendees: 50,
-      maxPrice: 10000,
-      name: "Donate",
-      unitPrice: 3000,
-    });
-    // Base £30, but the buyer chooses to pay £80 (within the £100 cap).
-    await runStripeSuccess({
-      email: "generous@example.com",
-      items: singleItem(listing.id, 1, 8000),
-      name: "Generous",
-      paymentIntent: "pi_more",
-      sessionId: "cs_more",
-      total: 8000,
-    });
-    const attendeeId = (await getAttendeesRaw(listing.id))[0]!.id;
-
-    expect(await incomeOf(listing.id)).toBe(8000);
-    expect(await owedBy(attendeeId)).toBe(0);
-    expect(await worldBalance()).toBe(-8000);
-    // Both income surfaces render the chosen £80 (no refund has touched it).
-    await assertRenderedIncome(listing.id, 8000);
-  });
-});
